Moved comparison code into a separate Comparer class.
